// data/github/current/ui/packages/issue-metadata/components/projects/fields/__tests__/__generated__/NumberFieldTestQuery.graphql.ts
export type ProjectV2FieldType = "ASSIGNEES" | "DATE" | "ISSUE_TYPE" | "ITERATION" | "LABELS" | "LINKED_PULL_REQUESTS" | "MILESTONE" | "NUMBER" | "PARENT_ISSUE" | "REPOSITORY" | "REVIEWERS" | "SINGLE_SELECT" | "SUB_ISSUES_PROGRESS" | "TEXT" | "TITLE" | "TRACKED_BY" | "TRACKS" | "%future added value";
export type NumberFieldTestQuery$variables = Record<PropertyKey, never>;
export type NumberFieldTestQuery$data = {
  readonly node: {
    readonly fieldValues?: {
      readonly edges: ReadonlyArray<{
        readonly node: {
          readonly field?: {
            readonly id?: string;
          };
          readonly id?: string;
          readonly " $fragmentSpreads": FragmentRefs<"NumberFieldFragment">;
        } | null | undefined;
      } | null | undefined> | null | undefined;
    };
    readonly project?: {
      readonly fields: {
        readonly edges: ReadonlyArray<{
          readonly node: {
            readonly dataType?: ProjectV2FieldType;
            readonly id?: string;
            readonly name?: string;
            readonly " $fragmentSpreads": FragmentRefs<"NumberFieldConfigFragment">;
          } | null | undefined;
        } | null | undefined> | null | undefined;
      };
      readonly id: string;
      readonly url: string;
      readonly viewerCanUpdate: boolean;
    };
  } | null | undefined;
};
export type NumberFieldTestQuery = {
  response: NumberFieldTestQuery$data;
  variables: NumberFieldTestQuery$variables;
};